Tests run on Django 2
