+++ runtests.py
@@ -1,53 +1,15 @@
 #!/usr/bin/env python
+import os
+import sys
+
 import django
 from django.conf import settings
-from django.core.management import call_command
+from django.test.utils import get_runner
 
-
-INSTALLED_APPS = (
-    # Required contrib apps.
-    'django.contrib.admin',
-    'django.contrib.auth',
-    'django.contrib.contenttypes',
-    'django.contrib.sites',
-    'django.contrib.sessions',
-    # Our app and it's test app.
-    'dbsettings',
-)
-
-SETTINGS = {
-    'INSTALLED_APPS': INSTALLED_APPS,
-    'SITE_ID': 1,
-    'ROOT_URLCONF': 'dbsettings.tests.test_urls',
-    'DATABASES': {
-        'default': {
-            'ENGINE': 'django.db.backends.sqlite3',
-            'NAME': ':memory:',
-        }
-    },
-    'MIDDLEWARE': (
-        'django.contrib.sessions.middleware.SessionMiddleware',
-        'django.contrib.auth.middleware.AuthenticationMiddleware',
-        'django.contrib.messages.middleware.MessageMiddleware',
-    ),
-    'MIGRATION_MODULES': {
-        # This allow test models to be created even if they are not in migration.
-        # Still no better solution available: https://code.djangoproject.com/ticket/7835
-        # This hack is used in Django testrunner itself.
-        'dbsettings': 'dbsettings.skip_migrations_for_test',
-    },
-    'TEMPLATES': [
-        {
-            'BACKEND': 'django.template.backends.django.DjangoTemplates',
-            'APP_DIRS': True,
-        },
-    ],
-}
-
-if not settings.configured:
-    settings.configure(**SETTINGS)
-
-if django.VERSION >= (1, 7):
-        django.setup()
-
-call_command('test', 'dbsettings')
+if __name__ == "__main__":
+    os.environ['DJANGO_SETTINGS_MODULE'] = 'tests.test_settings'
+    django.setup()
+    TestRunner = get_runner(settings)
+    test_runner = TestRunner()  # debug : verbosity=2, keepdb=True
+    failures = test_runner.run_tests(["tests"])
+    sys.exit(bool(failures))
